+++ src/main/resources/templates/admin/Users.html
@@ -72,7 +72,7 @@
                 "language": {
                     "url": "https://cdn.datatables.net/plug-ins/3cfcc339e89/i18n/Japanese.json"
                 },
-                "ajax": { url: /*[[@{/getuserlist}]]*/ 'getuserlist', dataSrc: '' },
+                "ajax": { url: /*[[@{/admin/search}]]*/ 'admin/search', dataSrc: '' },
                 "columns": [
                     { data: "userName" },
                     { data: "lastName"}
